Modularize database runtime tests
Separate sync-hook, activity, recovery, configuration, and open/status coverage into focused suites while preserving runtime behavior.
