+++ plugins/db/src/runtime.rs
@@ -2396,3208 +2396,4 @@ fn bind_params<'q>(
 }
 
 #[cfg(test)]
-mod tests {
-    use std::sync::atomic::{AtomicBool, Ordering};
-
-    use super::*;
-    use wiremock::{Mock, MockServer, Request, Respond, ResponseTemplate, matchers::path};
-
-    #[derive(Clone, Default)]
-    struct InitiallyUninitializedWitness {
-        initialized: std::sync::Arc<AtomicBool>,
-    }
-
-    impl Respond for InitiallyUninitializedWitness {
-        fn respond(&self, request: &Request) -> ResponseTemplate {
-            if request.method == wiremock::http::Method::POST {
-                let body: serde_json::Value = request.body_json().unwrap();
-                if !self.initialized.load(Ordering::SeqCst)
-                    && body["initialize"].as_bool() != Some(true)
-                {
-                    return ResponseTemplate::new(409);
-                }
-                if body["events"].as_array().is_none_or(Vec::is_empty) {
-                    return ResponseTemplate::new(400);
-                }
-                self.initialized.store(true, Ordering::SeqCst);
-                return ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                    "initializedAt": "2026-07-23T00:00:00Z",
-                    "headSequence": 0,
-                }));
-            }
-
-            let initialized = self.initialized.load(Ordering::SeqCst);
-            ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "initialized": initialized,
-                "initializedAt": initialized.then_some("2026-07-23T00:00:00Z"),
-                "headSequence": 0,
-                "throughSequence": 0,
-                "nextAfterSequence": 0,
-                "events": [],
-            }))
-        }
-    }
-
-    async fn configured_test_hook() -> (E2eeSyncHook, MockServer) {
-        let hook = E2eeSyncHook::default();
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        hook.set_personal_workspace("workspace-1", &recovery_key)
-            .unwrap();
-        let witness_state = InitiallyUninitializedWitness::default();
-        witness_state.initialized.store(true, Ordering::SeqCst);
-        let witness_server = MockServer::start().await;
-        Mock::given(path("/sync/e2ee/witness/workspace-1"))
-            .respond_with(witness_state)
-            .mount(&witness_server)
-            .await;
-        hook.set_witness(
-            crate::e2ee_witness::E2eeWitnessClient::new(
-                crate::CloudsyncE2eeWitness {
-                    endpoint: format!("{}/sync/e2ee/witness/workspace-1", witness_server.uri()),
-                    access_token: "access-token".to_string(),
-                },
-                "workspace-1",
-            )
-            .unwrap(),
-        );
-        (hook, witness_server)
-    }
-
-    fn receive_result(chunks: u64, complete: bool) -> anlg_db_core::CloudsyncNetworkResult {
-        serde_json::from_value(serde_json::json!({
-            "receive": {
-                "rows": chunks,
-                "tables": [],
-                "chunks": chunks,
-                "bytes": chunks,
-                "complete": complete
-            }
-        }))
-        .unwrap()
-    }
-
-    #[tokio::test]
-    async fn capture_lifecycle_marker_defers_only_its_transcript() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        let hook = E2eeSyncHook::default();
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        hook.set_personal_workspace("workspace-1", &recovery_key)
-            .unwrap();
-        let witness_state = InitiallyUninitializedWitness::default();
-        witness_state.initialized.store(true, Ordering::SeqCst);
-        let witness_server = MockServer::start().await;
-        Mock::given(path("/sync/e2ee/witness/workspace-1"))
-            .respond_with(witness_state)
-            .mount(&witness_server)
-            .await;
-        hook.set_witness(
-            crate::e2ee_witness::E2eeWitnessClient::new(
-                crate::CloudsyncE2eeWitness {
-                    endpoint: format!("{}/sync/e2ee/witness/workspace-1", witness_server.uri()),
-                    access_token: "access-token".to_string(),
-                },
-                "workspace-1",
-            )
-            .unwrap(),
-        );
-
-        sqlx::query(
-            "INSERT INTO app_settings (id, value_json)
-             VALUES ('capture_lifecycle_pending:session-1', ?)",
-        )
-        .bind(
-            serde_json::json!({
-                "version": 1,
-                "phase": "capturing",
-                "sessionId": "session-1",
-                "transcriptId": "transcript-1",
-                "startedAt": 1_000,
-                "createdAt": "2026-07-24T00:00:00.000Z",
-                "audioOffsetMs": 0,
-                "preserveExistingTranscript": false,
-                "ownerUserId": "workspace-1",
-                "memo": ""
-            })
-            .to_string(),
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, title)
-             VALUES
-               ('session-1', 'workspace-1', 'Active capture'),
-               ('session-2', 'workspace-1', 'Unrelated edit')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO transcripts (id, workspace_id, session_id, words_json)
-             VALUES
-               ('transcript-1', 'workspace-1', 'session-1', '[{\"text\":\"partial\"}]'),
-               ('transcript-2', 'workspace-1', 'session-2', '[{\"text\":\"complete\"}]')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        assert_eq!(
-            anlg_db_core::CloudsyncSyncHook::before_sync(&hook, db.pool())
-                .await
-                .unwrap(),
-            anlg_db_core::CloudsyncSyncDirective::SendAndReceive
-        );
-        assert!(!witness_server.received_requests().await.unwrap().is_empty());
-        let remaining_dirty: Vec<(String, String)> = sqlx::query_as(
-            "SELECT table_name, row_id
-             FROM e2ee_dirty_rows
-             ORDER BY table_name, row_id",
-        )
-        .fetch_all(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(
-            remaining_dirty,
-            vec![("transcripts".to_string(), "transcript-1".to_string())]
-        );
-        let protected_records: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*)
-             FROM e2ee_local_state
-             WHERE table_name = 'transcripts' AND row_id = 'transcript-1'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(protected_records, 0);
-        let unrelated_records: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*)
-             FROM e2ee_local_state
-             WHERE table_name = 'transcripts' AND row_id = 'transcript-2'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert!(unrelated_records > 0);
-
-        let result: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "receive": {
-                    "rows": 0,
-                    "tables": [],
-                    "chunks": 0,
-                    "bytes": 0,
-                    "complete": true
-                }
-            }))
-            .unwrap();
-        let outcome = anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &result)
-            .await
-            .unwrap();
-        assert!(!outcome.local_work_remaining);
-        let keys = hook.snapshot();
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(value_json, '$.phase', 'finalizing')
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "INSERT INTO app_settings (id, value_json)
-             VALUES
-               ('capture_lifecycle_pending:malformed', '{}'),
-               ('capture_lifecycle_pending:invalid-json', '{')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_remove(value_json, '$.phase')
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            !anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(value_json, '$.phase', 'unknown')
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            !anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(value_json, '$.summaryMode', 'if_empty')
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_remove(
-               json_set(value_json, '$.version', json('true')),
-               '$.summaryMode'
-             )
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(
-               value_json,
-               '$.version',
-               1,
-               '$.startedAt',
-               1e999
-             )
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            anlg_db_app::has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &keys,
-            )
-            .await
-            .unwrap()
-        );
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(
-               value_json,
-               '$.startedAt',
-               1000,
-               '$.summaryMode',
-               'if_empty'
-             )
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let outcome = anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &result)
-            .await
-            .unwrap();
-        assert!(outcome.local_work_remaining);
-
-        assert_eq!(
-            anlg_db_core::CloudsyncSyncHook::before_sync(&hook, db.pool())
-                .await
-                .unwrap(),
-            anlg_db_core::CloudsyncSyncDirective::SendAndReceive
-        );
-        let protected_dirty: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*)
-             FROM e2ee_dirty_rows
-             WHERE table_name = 'transcripts' AND row_id = 'transcript-1'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(protected_dirty, 0);
-    }
-
-    #[tokio::test]
-    async fn idle_sync_skips_replica_reconciliation_until_new_work_arrives() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        let (hook, _witness_server) = configured_test_hook().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-1', 'workspace-1', 'Local session')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        anlg_db_core::CloudsyncSyncHook::before_sync(&hook, db.pool())
-            .await
-            .unwrap();
-        let complete_idle = receive_result(0, true);
-        anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &complete_idle)
-            .await
-            .unwrap();
-
-        let (record_id, original_payload): (String, String) = sqlx::query_as(
-            "SELECT replica.id, replica.payload
-             FROM e2ee_records AS replica
-             INNER JOIN e2ee_witness_records AS witness
-               ON witness.workspace_id = replica.workspace_id
-              AND witness.record_id = replica.id
-              AND witness.payload = replica.payload
-             WHERE replica.workspace_id = 'workspace-1'
-             ORDER BY replica.id
-             LIMIT 1",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = 'corrupt' WHERE id = ?")
-            .bind(&record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-
-        let idle_outcome =
-            anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &complete_idle)
-                .await
-                .unwrap();
-        let idle_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(idle_payload, "corrupt");
-        assert!(!idle_outcome.local_work_remaining);
-
-        let delivered = receive_result(1, true);
-        anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &delivered)
-            .await
-            .unwrap();
-        let reconciled_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(reconciled_payload, original_payload);
-    }
-
-    #[tokio::test]
-    async fn nonfinal_receive_keeps_reconciliation_pending_for_final_snapshot() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        let (hook, _witness_server) = configured_test_hook().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-1', 'workspace-1', 'Remote session')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        anlg_db_core::CloudsyncSyncHook::before_sync(&hook, db.pool())
-            .await
-            .unwrap();
-        anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &receive_result(0, true))
-            .await
-            .unwrap();
-        let missing_record_id: String = sqlx::query_scalar(
-            "SELECT replica.id
-             FROM e2ee_records AS replica
-             INNER JOIN e2ee_witness_records AS witness
-               ON witness.workspace_id = replica.workspace_id
-              AND witness.record_id = replica.id
-             WHERE replica.workspace_id = 'workspace-1'
-             ORDER BY replica.id
-             LIMIT 1",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        sqlx::query("DELETE FROM e2ee_records WHERE id = ?")
-            .bind(&missing_record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-
-        let partial_outcome = anlg_db_core::CloudsyncSyncHook::after_sync(
-            &hook,
-            db.pool(),
-            &receive_result(1, false),
-        )
-        .await
-        .unwrap();
-        assert!(partial_outcome.local_work_remaining);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records WHERE id = ?")
-                .bind(&missing_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            0
-        );
-
-        let final_outcome =
-            anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &receive_result(0, true))
-                .await
-                .unwrap();
-        assert!(!final_outcome.local_work_remaining);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records WHERE id = ?")
-                .bind(&missing_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            1
-        );
-
-        sqlx::query("DELETE FROM e2ee_records WHERE id = ?")
-            .bind(&missing_record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let failed_after_chunks: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "receive": {
-                    "rows": 1,
-                    "tables": ["e2ee_records"],
-                    "chunks": 1,
-                    "bytes": 1,
-                    "complete": false,
-                    "error": "later chunk failed"
-                }
-            }))
-            .unwrap();
-        let failed_outcome =
-            anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &failed_after_chunks)
-                .await
-                .unwrap();
-        assert!(failed_outcome.local_work_remaining);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records WHERE id = ?")
-                .bind(&missing_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            0
-        );
-
-        let retry_outcome =
-            anlg_db_core::CloudsyncSyncHook::after_sync(&hook, db.pool(), &receive_result(0, true))
-                .await
-                .unwrap();
-        assert!(!retry_outcome.local_work_remaining);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records WHERE id = ?")
-                .bind(missing_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            1
-        );
-    }
-
-    #[tokio::test]
-    async fn activity_drains_large_no_mismatch_preflight_before_local_write() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(std::sync::Arc::clone(&db)));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        let (configured_hook, _witness_server) = configured_test_hook().await;
-        runtime
-            .e2ee_sync_hook
-            .set_witness(configured_hook.witness().unwrap());
-
-        sqlx::query(
-            "WITH RECURSIVE rows(id) AS (
-               SELECT 1
-               UNION ALL
-               SELECT id + 1 FROM rows WHERE id < 50000
-             )
-             INSERT INTO e2ee_records (id, workspace_id, payload)
-             SELECT printf('!unwitnessed-%06d', id), 'workspace-1', 'unwitnessed' FROM rows",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_local_state (
-               record_id, workspace_id, table_name, row_id, field_name, revision,
-               writer_id, value_tag, payload_hash, payload
-             )
-             SELECT
-               id,
-               workspace_id,
-               'sessions',
-               id,
-               '$row',
-               1,
-               'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-               'unchanged',
-               'unchanged',
-               payload
-             FROM e2ee_records
-             WHERE id LIKE '!unwitnessed-%'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let key = runtime.workspace_key("workspace-1").unwrap();
-        let mut transaction = db.pool().begin().await.unwrap();
-        for index in 0..32 {
-            let sealed = key
-                .seal_field(
-                    "workspace-1",
-                    "sessions",
-                    &format!("received-session-{index:02}"),
-                    "$row",
-                    "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                    1,
-                    false,
-                    serde_json::json!(true),
-                )
-                .unwrap();
-            sqlx::query(
-                "INSERT INTO e2ee_records (id, workspace_id, payload)
-                 VALUES (?, 'workspace-1', ?)",
-            )
-            .bind(&sealed.record_id)
-            .bind(&sealed.payload)
-            .execute(&mut *transaction)
-            .await
-            .unwrap();
-            sqlx::query(
-                "INSERT INTO e2ee_witness_records (
-                   workspace_id, record_id, revision, writer_id, payload_hash, payload, sequence
-                 ) VALUES (
-                   'workspace-1', ?, 1, 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa', ?, ?, ?
-                 )",
-            )
-            .bind(&sealed.record_id)
-            .bind(anlg_e2ee::payload_hash(&sealed.payload))
-            .bind(&sealed.payload)
-            .bind(i64::from(index) + 1)
-            .execute(&mut *transaction)
-            .await
-            .unwrap();
-        }
-        transaction.commit().await.unwrap();
-
-        let reconcile_runtime = std::sync::Arc::clone(&runtime);
-        let reconcile_db = std::sync::Arc::clone(&db);
-        let reconcile = tokio::spawn(async move {
-            anlg_db_core::CloudsyncSyncHook::after_sync(
-                reconcile_runtime.e2ee_sync_hook.as_ref(),
-                reconcile_db.pool(),
-                &receive_result(1, false),
-            )
-            .await
-        });
-        tokio::time::timeout(std::time::Duration::from_secs(1), async {
-            while runtime
-                .e2ee_sync_hook
-                .received_apply_cancellation_checks
-                .load(Ordering::Acquire)
-                < 70
-            {
-                assert!(
-                    !reconcile.is_finished(),
-                    "received-replica reconciliation finished before cancellation"
-                );
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("received-replica reconciliation did not enter bounded preflight");
-
-        let cancelled_at = std::time::Instant::now();
-        runtime
-            .e2ee_sync_hook
-            .begin_activity("capture".to_string(), "session-1".to_string());
-        let outcome = tokio::time::timeout(std::time::Duration::from_secs(2), reconcile)
-            .await
-            .expect("activity did not drain received-replica reconciliation")
-            .unwrap()
-            .unwrap();
-        assert!(outcome.deferred);
-        assert!(outcome.local_work_remaining);
-        assert!(cancelled_at.elapsed() < std::time::Duration::from_secs(2));
-
-        tokio::time::timeout(
-            std::time::Duration::from_millis(250),
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('local-after-received-cancel', 'workspace-1', 'workspace-1', 'Local')",
-            )
-            .execute(db.pool()),
-        )
-        .await
-        .expect("local write remained blocked after received-replica cancellation")
-        .unwrap();
-        assert!(runtime.e2ee_sync_hook.end_activity("capture", "session-1"));
-        runtime.e2ee_sync_hook.notify_activity_changed();
-    }
-
-    #[test]
-    fn reconciliation_epochs_preserve_newer_requests() {
-        let hook = E2eeSyncHook::default();
-        hook.request_reconciliation();
-        let first_epoch = hook.reconciliation_request_epoch().unwrap();
-
-        hook.request_reconciliation();
-        hook.complete_reconciliation(first_epoch);
-
-        assert!(hook.reconciliation_requested());
-        let second_epoch = hook.reconciliation_request_epoch().unwrap();
-        assert!(second_epoch > first_epoch);
-        hook.complete_reconciliation(second_epoch);
-        assert!(!hook.reconciliation_requested());
-
-        hook.request_reconciliation();
-        hook.clear();
-        assert!(!hook.reconciliation_requested());
-        hook.request_reconciliation();
-        assert!(hook.reconciliation_requested());
-    }
-
-    #[tokio::test]
-    async fn witness_refresh_failure_keeps_reconciliation_pending() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        let hook = E2eeSyncHook::default();
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        hook.set_personal_workspace("workspace-1", &recovery_key)
-            .unwrap();
-        let witness_server = MockServer::start().await;
-        Mock::given(path("/sync/e2ee/witness/workspace-1"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "initialized": true,
-                "initializedAt": "2026-07-24T00:00:00Z",
-                "headSequence": 1,
-                "throughSequence": 1,
-                "nextAfterSequence": 1,
-                "events": [{
-                    "sequence": 1,
-                    "recordId": "invalid-record",
-                    "payloadHash": "invalid-hash",
-                    "payload": "invalid-payload"
-                }]
-            })))
-            .mount(&witness_server)
-            .await;
-        hook.set_witness(
-            crate::e2ee_witness::E2eeWitnessClient::new(
-                crate::CloudsyncE2eeWitness {
-                    endpoint: format!("{}/sync/e2ee/witness/workspace-1", witness_server.uri()),
-                    access_token: "access-token".to_string(),
-                },
-                "workspace-1",
-            )
-            .unwrap(),
-        );
-        let setup_epoch = hook.reconciliation_request_epoch().unwrap();
-        hook.complete_reconciliation(setup_epoch);
-
-        assert!(
-            anlg_db_core::CloudsyncSyncHook::before_sync(&hook, db.pool())
-                .await
-                .is_err()
-        );
-        assert!(hook.reconciliation_requested());
-    }
-
-    #[tokio::test]
-    async fn failed_reconciliation_epoch_stays_pending() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        let (hook, _witness_server) = configured_test_hook().await;
-        let setup_epoch = hook.reconciliation_request_epoch().unwrap();
-        hook.complete_reconciliation(setup_epoch);
-        sqlx::query(
-            "INSERT INTO e2ee_witness_records (
-               workspace_id,
-               record_id,
-               revision,
-               writer_id,
-               payload_hash,
-               payload,
-               sequence
-             )
-             VALUES (
-               'workspace-1',
-               'invalid-record',
-               1,
-               'invalid-writer',
-               'invalid-hash',
-               'invalid-payload',
-               1
-             )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        assert!(
-            anlg_db_core::CloudsyncSyncHook::after_sync(
-                &hook,
-                db.pool(),
-                &receive_result(1, true),
-            )
-            .await
-            .is_err()
-        );
-        assert!(hook.reconciliation_requested());
-    }
-
-    #[tokio::test]
-    async fn cloudsync_activity_leases_are_idempotent_and_preserved_by_identity_clear() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-2".to_string())
-            .await
-            .unwrap();
-        runtime.e2ee_sync_hook.clear();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["activity_paused"], true);
-        assert_eq!(status["deferred_for_capture"], true);
-        assert_eq!(status["has_unsent_changes"], serde_json::Value::Null);
-
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "missing".to_string())
-            .await
-            .unwrap();
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-2".to_string())
-            .await
-            .unwrap();
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        assert!(!runtime.e2ee_sync_hook.activity_paused());
-    }
-
-    #[tokio::test]
-    async fn cloudsync_activity_pause_is_process_scoped() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        drop(runtime);
-
-        let restarted_runtime = PluginDbRuntime::new(db);
-        let status = restarted_runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["activity_paused"], false);
-        assert_eq!(status["deferred_for_capture"], false);
-    }
-
-    #[tokio::test]
-    async fn activity_resume_wait_cannot_lose_its_notification() {
-        let hook = E2eeSyncHook::default();
-        hook.begin_activity("capture".to_string(), "session-1".to_string());
-        let mut resumed = Box::pin(hook.wait_until_activity_resumed());
-
-        tokio::select! {
-            biased;
-            () = &mut resumed => panic!("activity resume wait completed while paused"),
-            _ = tokio::task::yield_now() => {}
-        }
-        assert!(hook.end_activity("capture", "session-1"));
-        hook.notify_activity_changed();
-        tokio::time::timeout(std::time::Duration::from_millis(100), resumed)
-            .await
-            .expect("activity resume notification was lost");
-
-        hook.wait_until_activity_resumed().await;
-    }
-
-    #[tokio::test]
-    async fn clearing_activities_wakes_resume_waiters() {
-        let hook = E2eeSyncHook::default();
-        hook.begin_activity("capture".to_string(), "session-1".to_string());
-        hook.begin_activity("chat".to_string(), "chat-1".to_string());
-        let mut resumed = Box::pin(hook.wait_until_activity_resumed());
-
-        tokio::select! {
-            biased;
-            () = &mut resumed => panic!("activity resume wait completed while paused"),
-            _ = tokio::task::yield_now() => {}
-        }
-        hook.clear_activities();
-        tokio::time::timeout(std::time::Duration::from_millis(100), resumed)
-            .await
-            .expect("activity clear notification was lost");
-        assert!(!hook.activity_paused());
-    }
-
-    #[tokio::test]
-    async fn paused_cloudsync_status_does_not_probe_the_busy_database() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        let _held_connection = db.pool().acquire().await.unwrap();
-
-        let status = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.cloudsync_status(),
-        )
-        .await
-        .expect("paused CloudSync status queried the busy database")
-        .unwrap();
-
-        assert_eq!(status["activity_paused"], true);
-        assert_eq!(status["deferred_for_capture"], true);
-    }
-
-    #[tokio::test]
-    async fn activity_begin_waits_for_an_in_flight_cloudsync_control_operation() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let control_operation = runtime.cloudsync_control_operation.lock().await;
-        let mut begin = Box::pin(
-            runtime.begin_cloudsync_activity("capture".to_string(), "session-1".to_string()),
-        );
-
-        tokio::select! {
-            biased;
-            result = &mut begin => panic!("activity begin bypassed recovery: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-
-        drop(control_operation);
-        tokio::time::timeout(std::time::Duration::from_millis(100), begin)
-            .await
-            .expect("activity begin did not finish after CloudSync control became idle")
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn activity_begin_timeout_fails_closed_and_releases_the_new_lease() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let control_operation = runtime.cloudsync_control_operation.lock().await;
-
-        let error = tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            runtime.begin_cloudsync_activity_with_timeout(
-                "capture".to_string(),
-                "session-1".to_string(),
-                std::time::Duration::from_millis(10),
-            ),
-        )
-        .await
-        .expect("activity begin exceeded its drain deadline")
-        .unwrap_err();
-
-        assert!(matches!(error, crate::Error::CloudsyncActivityDrainTimeout));
-        assert!(!runtime.e2ee_sync_hook.activity_paused());
-        assert!(
-            !runtime
-                .e2ee_sync_hook
-                .has_activity_lease("capture", "session-1")
-        );
-        drop(control_operation);
-    }
-
-    #[tokio::test]
-    async fn duplicate_activity_begin_timeout_preserves_the_existing_lease() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        let control_operation = runtime.cloudsync_control_operation.lock().await;
-
-        runtime
-            .begin_cloudsync_activity_with_timeout(
-                "capture".to_string(),
-                "session-1".to_string(),
-                std::time::Duration::from_millis(10),
-            )
-            .await
-            .unwrap();
-
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        assert!(
-            runtime
-                .e2ee_sync_hook
-                .has_activity_lease("capture", "session-1")
-        );
-        drop(control_operation);
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn activity_end_cancels_a_begin_waiting_for_cloudsync_control() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let control_operation = runtime.cloudsync_control_operation.lock().await;
-        let mut begin = Box::pin(
-            runtime.begin_cloudsync_activity("capture".to_string(), "session-1".to_string()),
-        );
-
-        tokio::select! {
-            biased;
-            result = &mut begin => panic!("activity begin bypassed CloudSync control: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        drop(control_operation);
-
-        let error = tokio::time::timeout(std::time::Duration::from_millis(100), begin)
-            .await
-            .expect("cancelled activity begin did not finish")
-            .unwrap_err();
-        assert!(
-            error
-                .to_string()
-                .contains("ended before synchronization became idle")
-        );
-    }
-
-    #[tokio::test]
-    async fn cloudsync_configuration_and_start_fail_promptly_during_activity() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        let config = serde_json::to_string(&anlg_db_core::CloudsyncRuntimeConfig {
-            connection_string: "managed-database-id".to_string(),
-            auth: anlg_db_core::CloudsyncAuth::None,
-            tables: Vec::new(),
-            sync_interval_ms: DEFAULT_CLOUDSYNC_INTERVAL_MS,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        })
-        .unwrap();
-
-        let configure_error = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.configure_cloudsync(config),
-        )
-        .await
-        .expect("CloudSync configuration waited for the activity to finish")
-        .unwrap_err();
-        assert!(matches!(
-            configure_error,
-            crate::Error::CloudsyncActivityDeferred
-        ));
-        assert_eq!(configure_error.to_string(), "cloudsync_activity_deferred");
-
-        let token_error = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.configure_cloudsync_token(
-                "managed-database-id".to_string(),
-                "token".to_string(),
-                "user-a".to_string(),
-                crate::CloudsyncE2eeWitness {
-                    endpoint: "https://example.com/witness".to_string(),
-                    access_token: "access-token".to_string(),
-                },
-            ),
-        )
-        .await
-        .expect("CloudSync token configuration waited for the activity to finish")
-        .unwrap_err();
-        assert!(matches!(
-            token_error,
-            crate::Error::CloudsyncActivityDeferred
-        ));
-
-        let start_error = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.start_cloudsync(),
-        )
-        .await
-        .expect("CloudSync start waited for the activity to finish")
-        .unwrap_err();
-        assert!(matches!(
-            start_error,
-            crate::Error::CloudsyncActivityDeferred
-        ));
-    }
-
-    #[tokio::test]
-    async fn cloudsync_start_prearms_reconciliation() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        sqlx::query(
-            "UPDATE storage_migration_state
-             SET importer_version = ?, parity_verified = 1
-             WHERE id = 'legacy_v1'",
-        )
-        .bind(anlg_db_app::LEGACY_IMPORTER_VERSION)
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        let setup_epoch = runtime
-            .e2ee_sync_hook
-            .reconciliation_request_epoch()
-            .unwrap();
-        runtime.e2ee_sync_hook.complete_reconciliation(setup_epoch);
-
-        let _ = runtime.start_cloudsync().await;
-
-        assert!(runtime.e2ee_sync_hook.reconciliation_requested());
-    }
-
-    #[tokio::test]
-    async fn cloudsync_control_rechecks_activity_after_acquiring_serialization() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let control_operation = runtime.cloudsync_control_operation.lock().await;
-        let mut start = Box::pin(runtime.start_cloudsync());
-
-        tokio::select! {
-            biased;
-            result = &mut start => panic!("CloudSync start bypassed control serialization: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-        runtime
-            .e2ee_sync_hook
-            .begin_activity("capture".to_string(), "session-1".to_string());
-
-        let error = tokio::time::timeout(std::time::Duration::from_millis(100), start)
-            .await
-            .expect("CloudSync start remained queued after activity began")
-            .unwrap_err();
-        assert!(matches!(error, crate::Error::CloudsyncActivityDeferred));
-        drop(control_operation);
-    }
-
-    #[tokio::test]
-    async fn local_account_binding_is_not_deferred_by_cloudsync_activity() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-
-        let bound = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.bind_cloudsync_account("user-a".to_string()),
-        )
-        .await
-        .expect("local account binding waited for the activity to finish")
-        .unwrap();
-        assert!(bound);
-        assert!(
-            runtime
-                .e2ee_sync_hook
-                .has_activity_lease("capture", "session-1")
-        );
-    }
-
-    #[tokio::test]
-    async fn local_account_binding_times_out_before_mutation_and_remains_fail_closed() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let control = runtime.cloudsync_control_operation.lock().await;
-
-        let error = runtime
-            .bind_cloudsync_account_with_lock_timeout(
-                "user-a".to_string(),
-                std::time::Duration::from_millis(10),
-            )
-            .await
-            .unwrap_err();
-        assert_eq!(
-            error.to_string(),
-            "CloudSync account binding lock preflight timed out"
-        );
-        drop(control);
-
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-        assert!(
-            !runtime
-                .bind_cloudsync_account("user-b".to_string())
-                .await
-                .unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn stop_suspend_and_logout_are_not_deferred_by_cloudsync_activity() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-
-        let stop = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.stop_cloudsync(),
-        )
-        .await
-        .expect("CloudSync stop waited for the activity to finish");
-        assert!(!matches!(
-            stop,
-            Err(crate::Error::CloudsyncActivityDeferred)
-        ));
-
-        let suspend = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.suspend_cloudsync(),
-        )
-        .await
-        .expect("CloudSync suspension waited for the activity to finish");
-        assert!(!matches!(
-            suspend,
-            Err(crate::Error::CloudsyncActivityDeferred)
-        ));
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-
-        let logout = tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.logout_cloudsync(false),
-        )
-        .await
-        .expect("CloudSync logout waited for the activity to finish");
-        assert!(!matches!(
-            logout,
-            Err(crate::Error::CloudsyncActivityDeferred)
-        ));
-        assert!(!runtime.e2ee_sync_hook.activity_paused());
-    }
-
-    #[tokio::test]
-    async fn sign_out_suspend_preserves_activity_leases() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        runtime
-            .begin_cloudsync_activity("chat".to_string(), "chat-1".to_string())
-            .await
-            .unwrap();
-
-        runtime.suspend_cloudsync_for_sign_out().await.unwrap();
-
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        assert!(
-            runtime
-                .e2ee_sync_hook
-                .has_activity_lease("capture", "session-1")
-        );
-        assert!(runtime.e2ee_sync_hook.has_activity_lease("chat", "chat-1"));
-    }
-
-    #[tokio::test]
-    async fn sign_out_suspend_preserves_activity_leases_after_non_busy_error() {
-        let db = std::sync::Arc::new(Db::connect_memory().await.unwrap());
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        db.pool().close().await;
-
-        let error = runtime.suspend_cloudsync_for_sign_out().await.unwrap_err();
-
-        assert!(!matches!(
-            error,
-            crate::Error::Cloudsync(anlg_db_core::CloudsyncRuntimeError::LocalStatusBusy)
-        ));
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        assert!(!runtime.e2ee_sync_hook.activity_paused());
-    }
-
-    #[tokio::test]
-    async fn auth_loss_suspend_preserves_activity_leases_after_non_busy_error() {
-        let db = std::sync::Arc::new(Db::connect_memory().await.unwrap());
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        db.pool().close().await;
-
-        let error = runtime
-            .suspend_cloudsync_after_auth_loss()
-            .await
-            .unwrap_err();
-
-        assert!(!matches!(
-            error,
-            crate::Error::Cloudsync(anlg_db_core::CloudsyncRuntimeError::LocalStatusBusy)
-        ));
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        assert!(!runtime.e2ee_sync_hook.activity_paused());
-    }
-
-    #[tokio::test]
-    async fn auth_suspension_bypasses_activity_acquisition_without_clearing_leases() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .e2ee_sync_hook
-            .begin_activity("capture".to_string(), "session-1".to_string());
-        let acquisition = runtime.cloudsync_activity_acquisition.lock().await;
-        tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.suspend_cloudsync_after_auth_loss(),
-        )
-        .await
-        .expect("auth-loss suspension waited for activity acquisition")
-        .unwrap();
-        tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.suspend_cloudsync_for_sign_out(),
-        )
-        .await
-        .expect("sign-out suspension waited for activity acquisition")
-        .unwrap();
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        drop(acquisition);
-    }
-
-    #[tokio::test]
-    async fn raw_suspend_bypasses_activity_acquisition_and_preserves_leases() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .e2ee_sync_hook
-            .begin_activity("chat".to_string(), "chat-1".to_string());
-        let acquisition = runtime.cloudsync_activity_acquisition.lock().await;
-        let mut queued_activity = Box::pin(
-            runtime.begin_cloudsync_activity("capture".to_string(), "session-1".to_string()),
-        );
-        tokio::select! {
-            biased;
-            result = &mut queued_activity => panic!("queued activity unexpectedly completed: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        tokio::time::timeout(
-            std::time::Duration::from_millis(100),
-            runtime.suspend_cloudsync(),
-        )
-        .await
-        .expect("forced auth cleanup waited for activity acquisition")
-        .unwrap();
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        drop(acquisition);
-        tokio::time::timeout(std::time::Duration::from_millis(100), queued_activity)
-            .await
-            .expect("queued activity remained blocked")
-            .unwrap();
-        assert!(
-            runtime
-                .e2ee_sync_hook
-                .has_activity_lease("capture", "session-1")
-        );
-        assert!(runtime.e2ee_sync_hook.has_activity_lease("chat", "chat-1"));
-    }
-
-    #[tokio::test]
-    async fn final_activity_release_resets_the_recovery_delay_clock() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        {
-            let mut schedule = runtime.scheduled_cloudsync_full_resync.lock().unwrap();
-            schedule.claim("generation-1");
-            schedule.last_progress_at =
-                Some(std::time::Instant::now() - std::time::Duration::from_secs(61));
-        }
-
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-
-        let schedule = runtime.scheduled_cloudsync_full_resync.lock().unwrap();
-        assert!(!schedule.recovery_delayed);
-        assert!(!schedule.is_delayed("generation-1"));
-    }
-
-    #[tokio::test]
-    async fn activity_release_does_not_hide_a_recovery_failure() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        runtime
-            .begin_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        {
-            let mut schedule = runtime.scheduled_cloudsync_full_resync.lock().unwrap();
-            schedule.claim("generation-1");
-            schedule.mark_failure("generation-1");
-        }
-
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-
-        let schedule = runtime.scheduled_cloudsync_full_resync.lock().unwrap();
-        assert!(schedule.recovery_delayed);
-        assert!(schedule.is_delayed("generation-1"));
-    }
-
-    #[test]
-    fn cloudsync_activity_identifiers_are_bounded() {
-        assert!(normalized_cloudsync_activity_part(" ".to_string(), "activity").is_err());
-        assert!(
-            normalized_cloudsync_activity_part("capture\n".to_string(), "activity").is_ok(),
-            "outer whitespace is intentionally trimmed"
-        );
-        assert!(normalized_cloudsync_activity_part("cap\nture".to_string(), "activity").is_err());
-        assert!(normalized_cloudsync_activity_part("a".repeat(33), "activity").is_err());
-        assert!(normalized_cloudsync_activity_part("k".repeat(129), "key").is_err());
-    }
-
-    #[test]
-    fn recovery_waits_longer_without_progress() {
-        assert_eq!(
-            cloudsync_recovery_step_delay(CloudsyncRecoveryStep::Progressed),
-            Some(CLOUDSYNC_FULL_RESYNC_PROGRESS_INTERVAL)
-        );
-        assert_eq!(
-            cloudsync_recovery_step_delay(CloudsyncRecoveryStep::Waiting),
-            Some(CLOUDSYNC_FULL_RESYNC_RETRY_INTERVAL)
-        );
-        assert_eq!(
-            cloudsync_recovery_step_delay(CloudsyncRecoveryStep::Complete),
-            None
-        );
-    }
-
-    #[test]
-    fn cloudsync_completion_requires_confirmed_send_and_receive() {
-        let completed: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "send": {
-                    "status": "synced",
-                    "localVersion": 2,
-                    "serverVersion": 2
-                },
-                "receive": {
-                    "rows": 0,
-                    "tables": [],
-                    "complete": true
-                }
-            }))
-            .unwrap();
-        let unconfirmed_send: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "send": {
-                    "status": "syncing",
-                    "localVersion": 2,
-                    "serverVersion": 1
-                }
-            }))
-            .unwrap();
-        let incomplete_receive: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "send": {
-                    "status": "synced",
-                    "localVersion": 2,
-                    "serverVersion": 2
-                },
-                "receive": {
-                    "rows": 1,
-                    "tables": ["sessions"],
-                    "complete": false
-                }
-            }))
-            .unwrap();
-        let send_only: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "send": {
-                    "status": "synced",
-                    "localVersion": 2,
-                    "serverVersion": 2
-                }
-            }))
-            .unwrap();
-        let receive_only: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "receive": {
-                    "rows": 0,
-                    "tables": [],
-                    "complete": true
-                }
-            }))
-            .unwrap();
-        let delivered_final: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "receive": {
-                    "rows": 1,
-                    "tables": ["e2ee_records"],
-                    "chunks": 1,
-                    "complete": true
-                }
-            }))
-            .unwrap();
-        let delivered_non_final: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "receive": {
-                    "rows": 1,
-                    "tables": ["e2ee_records"],
-                    "chunks": 1,
-                    "complete": false
-                }
-            }))
-            .unwrap();
-        let failed_receive: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "send": {
-                    "status": "synced",
-                    "localVersion": 2,
-                    "serverVersion": 2
-                },
-                "receive": {
-                    "rows": 0,
-                    "tables": [],
-                    "complete": true,
-                    "error": "database is locked"
-                }
-            }))
-            .unwrap();
-        let failed_after_chunks: anlg_db_core::CloudsyncNetworkResult =
-            serde_json::from_value(serde_json::json!({
-                "receive": {
-                    "rows": 1,
-                    "tables": ["e2ee_records"],
-                    "chunks": 1,
-                    "complete": false,
-                    "error": "later chunk failed"
-                }
-            }))
-            .unwrap();
-
-        assert!(cloudsync_send_completed(&completed));
-        assert!(cloudsync_receive_completed(&completed));
-        assert!(!cloudsync_send_completed(&unconfirmed_send));
-        assert!(!cloudsync_receive_completed(&incomplete_receive));
-        assert!(cloudsync_send_completed(&incomplete_receive));
-        assert!(cloudsync_send_completed(&send_only));
-        assert!(!cloudsync_receive_completed(&send_only));
-        assert!(!cloudsync_send_completed(&receive_only));
-        assert!(cloudsync_receive_completed(&receive_only));
-        assert!(!cloudsync_receive_delivered_final(&receive_only));
-        assert!(cloudsync_receive_delivered_final(&delivered_final));
-        assert!(cloudsync_receive_delivered(&delivered_non_final));
-        assert!(!cloudsync_recovery_snapshot_ready(
-            true,
-            &delivered_non_final
-        ));
-        assert!(cloudsync_recovery_snapshot_ready(true, &delivered_final));
-        assert!(!cloudsync_recovery_snapshot_ready(true, &receive_only));
-        assert!(!cloudsync_recovery_snapshot_ready(false, &delivered_final));
-        assert!(cloudsync_send_completed(&failed_receive));
-        assert!(!cloudsync_receive_completed(&failed_receive));
-        assert!(!cloudsync_receive_delivered(&failed_after_chunks));
-        assert!(cloudsync_receive_requires_reconciliation(
-            &failed_after_chunks
-        ));
-        assert!(!cloudsync_send_completed(
-            &anlg_db_core::CloudsyncNetworkResult::default()
-        ));
-    }
-
-    fn test_full_resync_config(token: &str) -> anlg_db_core::CloudsyncRuntimeConfig {
-        anlg_db_core::CloudsyncRuntimeConfig {
-            connection_string: "test-database".to_string(),
-            auth: anlg_db_core::CloudsyncAuth::Token {
-                token: token.to_string(),
-            },
-            tables: vec![],
-            sync_interval_ms: DEFAULT_CLOUDSYNC_INTERVAL_MS,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        }
-    }
-
-    #[test]
-    fn full_resync_schedule_tracks_generation_until_cancelled() {
-        let mut schedule = CloudsyncFullResyncSchedule::default();
-
-        schedule.claim("generation-1");
-        assert!(!schedule.is_delayed("generation-1"));
-        schedule.last_progress_at =
-            Some(std::time::Instant::now() - CLOUDSYNC_RECOVERY_DELAYED_AFTER);
-        assert!(schedule.is_delayed("generation-1"));
-        schedule.mark_progress("generation-1");
-        assert!(!schedule.is_delayed("generation-1"));
-        schedule.mark_failure("generation-1");
-        assert!(schedule.is_delayed("generation-1"));
-        schedule.mark_progress("generation-1");
-        assert!(!schedule.is_delayed("generation-1"));
-        schedule.claim("generation-1");
-        assert!(schedule.is_active("generation-1"));
-
-        schedule.cancel();
-        assert!(!schedule.is_active("generation-1"));
-    }
-
-    #[cfg(target_os = "macos")]
-    #[tokio::test]
-    async fn activity_drains_stalled_full_resync_before_immediate_local_write() {
-        use std::io::{Read, Write};
-
-        let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-        let endpoint = format!("http://{}", listener.local_addr().unwrap());
-        let (stalled_tx, stalled_rx) = std::sync::mpsc::sync_channel(1);
-        let (release_tx, release_rx) = std::sync::mpsc::sync_channel(1);
-        let server = std::thread::spawn(move || {
-            let (mut status_stream, _) = listener.accept().unwrap();
-            let mut request = [0_u8; 4096];
-            status_stream
-                .set_read_timeout(Some(std::time::Duration::from_secs(2)))
-                .unwrap();
-            let size = status_stream.read(&mut request).unwrap();
-            assert!(
-                String::from_utf8_lossy(&request[..size]).contains("/status"),
-                "recovery did not probe status before sending"
-            );
-            let body = r#"{"lastOptimisticVersion":0,"lastConfirmedVersion":0,"gaps":[]}"#;
-            write!(
-                status_stream,
-                "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-                body.len(),
-                body
-            )
-            .unwrap();
-            status_stream.flush().unwrap();
-            drop(status_stream);
-
-            let (stalled_stream, _) = listener.accept().unwrap();
-            stalled_tx.send(()).unwrap();
-            let _stalled_stream = stalled_stream;
-            let _ = release_rx.recv_timeout(std::time::Duration::from_secs(5));
-        });
-
-        let db = std::sync::Arc::new(Db::connect_memory().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        anlg_db_app::claim_cloudsync_workspace(db.pool(), "workspace-1")
-            .await
-            .unwrap();
-        db.cloudsync_init(CLOUDSYNC_REPLICA_TABLE, None, None)
-            .await
-            .unwrap();
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(std::sync::Arc::clone(&db)));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        let key = runtime.workspace_key("workspace-1").unwrap();
-        let (configured_hook, _witness_server) = configured_test_hook().await;
-        runtime
-            .e2ee_sync_hook
-            .set_witness(configured_hook.witness().unwrap());
-
-        let generation =
-            anlg_db_app::stage_cloudsync_poison_recovery(db.pool(), "workspace-1", "workspace-1")
-                .await
-                .unwrap();
-        let state = anlg_db_app::ensure_cloudsync_recovery_state(
-            db.pool(),
-            &generation,
-            "workspace-1",
-            "workspace-1",
-            &key,
-        )
-        .await
-        .unwrap();
-        assert!(
-            anlg_db_app::advance_cloudsync_recovery_phase(
-                db.pool(),
-                &generation,
-                anlg_db_app::CloudsyncRecoveryPhase::NeedFirstLogout,
-                anlg_db_app::CloudsyncRecoveryPhase::NeedBarrierInsert,
-            )
-            .await
-            .unwrap()
-        );
-        assert!(
-            anlg_db_app::insert_cloudsync_recovery_barrier(db.pool(), &state, &key)
-                .await
-                .unwrap()
-        );
-        assert!(
-            anlg_db_app::advance_cloudsync_recovery_phase(
-                db.pool(),
-                &generation,
-                anlg_db_app::CloudsyncRecoveryPhase::NeedBarrierInsert,
-                anlg_db_app::CloudsyncRecoveryPhase::NeedBarrierConfirm,
-            )
-            .await
-            .unwrap()
-        );
-
-        let config = anlg_db_core::CloudsyncRuntimeConfig {
-            connection_string: "test-database".to_string(),
-            auth: anlg_db_core::CloudsyncAuth::None,
-            tables: vec![],
-            sync_interval_ms: DEFAULT_CLOUDSYNC_INTERVAL_MS,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        };
-        db.cloudsync_prepare_manual_transport(config.clone())
-            .await
-            .unwrap();
-        sqlx::query("SELECT cloudsync_network_init_custom(?, ?)")
-            .bind(endpoint)
-            .bind("full-resync-interrupt-test")
-            .fetch_optional(db.pool())
-            .await
-            .unwrap();
-        let auth_generation = runtime.cloudsync_auth_generation();
-        runtime
-            .schedule_cloudsync_full_resync(generation.clone(), config, auth_generation)
-            .await;
-        tokio::task::spawn_blocking(move || {
-            stalled_rx
-                .recv_timeout(std::time::Duration::from_secs(5))
-                .expect("full resync did not reach the stalled manual send");
-        })
-        .await
-        .unwrap();
-
-        tokio::time::timeout(
-            std::time::Duration::from_millis(2_500),
-            runtime.begin_cloudsync_activity_with_timeout(
-                "capture".to_string(),
-                "session-1".to_string(),
-                std::time::Duration::from_millis(1_800),
-            ),
-        )
-        .await
-        .expect("activity did not drain the stalled full resync")
-        .unwrap();
-        tokio::time::timeout(
-            std::time::Duration::from_millis(250),
-            runtime.execute(
-                "INSERT INTO sessions (id, workspace_id, title)
-                 VALUES ('local-after-recovery', 'workspace-1', 'Local')"
-                    .to_string(),
-                vec![],
-            ),
-        )
-        .await
-        .expect("local write remained blocked after full-resync cancellation")
-        .unwrap();
-
-        assert_eq!(
-            anlg_db_app::cloudsync_recovery_state(db.pool())
-                .await
-                .unwrap()
-                .unwrap()
-                .phase,
-            anlg_db_app::CloudsyncRecoveryPhase::NeedBarrierConfirm
-        );
-        assert!(
-            runtime
-                .scheduled_cloudsync_full_resync
-                .lock()
-                .unwrap()
-                .is_active(&generation)
-        );
-
-        runtime.cancel_cloudsync_full_resync().await;
-        runtime
-            .end_cloudsync_activity("capture".to_string(), "session-1".to_string())
-            .await
-            .unwrap();
-        let _ = release_tx.send(());
-        server.join().unwrap();
-    }
-
-    async fn install_waiting_full_resync_task(
-        runtime: &PluginDbRuntime,
-    ) -> tokio::sync::oneshot::Receiver<()> {
-        let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
-        let (finished_tx, finished_rx) = tokio::sync::oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            let _ = shutdown_rx.await;
-            let _ = finished_tx.send(());
-        });
-        runtime
-            .scheduled_cloudsync_full_resync
-            .lock()
-            .unwrap()
-            .claim("test-generation");
-        *runtime.cloudsync_full_resync_task.lock().await = Some(CloudsyncFullResyncTask {
-            config: test_full_resync_config("test-token"),
-            generation: "test-generation".to_string(),
-            shutdown_tx: Some(shutdown_tx),
-            join_handle,
-        });
-        finished_rx
-    }
-
-    async fn install_blocked_full_resync_cancellation(
-        runtime: &PluginDbRuntime,
-    ) -> (
-        tokio::sync::oneshot::Sender<()>,
-        tokio::sync::oneshot::Receiver<()>,
-    ) {
-        let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
-        let (cancellation_started_tx, cancellation_started_rx) = tokio::sync::oneshot::channel();
-        let (release_tx, release_rx) = tokio::sync::oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            let _ = shutdown_rx.await;
-            let _ = cancellation_started_tx.send(());
-            let _ = release_rx.await;
-        });
-        runtime
-            .scheduled_cloudsync_full_resync
-            .lock()
-            .unwrap()
-            .claim("test-generation");
-        *runtime.cloudsync_full_resync_task.lock().await = Some(CloudsyncFullResyncTask {
-            config: test_full_resync_config("test-token"),
-            generation: "test-generation".to_string(),
-            shutdown_tx: Some(shutdown_tx),
-            join_handle,
-        });
-        (release_tx, cancellation_started_rx)
-    }
-
-    async fn assert_auth_invalidation_cancels_inflight_configuration(logout: bool) {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(db));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("user-a", &recovery_key)
-            .unwrap();
-        let (release_cancellation, cancellation_started) =
-            install_blocked_full_resync_cancellation(runtime.as_ref()).await;
-
-        let configure_runtime = std::sync::Arc::clone(&runtime);
-        let configure = tokio::spawn(async move {
-            configure_runtime
-                .configure_cloudsync_token(
-                    "managed-database-id".to_string(),
-                    "stale-token".to_string(),
-                    "user-a".to_string(),
-                    crate::CloudsyncE2eeWitness {
-                        endpoint: "https://example.com/witness".to_string(),
-                        access_token: "access-token".to_string(),
-                    },
-                )
-                .await
-        });
-        tokio::time::timeout(std::time::Duration::from_secs(1), cancellation_started)
-            .await
-            .expect("token configuration did not enter full-resync cancellation")
-            .unwrap();
-        let configuration_generation = runtime.cloudsync_auth_generation();
-
-        let invalidation_runtime = std::sync::Arc::clone(&runtime);
-        let invalidation = tokio::spawn(async move {
-            if logout {
-                invalidation_runtime.logout_cloudsync(false).await
-            } else {
-                invalidation_runtime.suspend_cloudsync().await
-            }
-        });
-        tokio::time::timeout(std::time::Duration::from_secs(1), async {
-            while runtime.cloudsync_auth_generation() == configuration_generation {
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("auth generation was not invalidated before waiting");
-
-        release_cancellation
-            .send(())
-            .expect("failed to release full-resync cancellation");
-        let error = tokio::time::timeout(std::time::Duration::from_secs(1), configure)
-            .await
-            .expect("cancelled token configuration did not finish")
-            .unwrap()
-            .unwrap_err();
-        assert!(matches!(
-            error,
-            crate::Error::CloudsyncConfigurationCancelled
-        ));
-        tokio::time::timeout(std::time::Duration::from_secs(1), invalidation)
-            .await
-            .expect("CloudSync invalidation did not finish")
-            .unwrap()
-            .unwrap();
-
-        assert!(runtime.workspace_key("user-a").is_none());
-        assert!(runtime.cloudsync_full_resync_task.lock().await.is_none());
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["configured"], false);
-        assert_eq!(status["network_initialized"], false);
-    }
-
-    #[tokio::test]
-    async fn suspend_invalidates_and_fails_closed_an_inflight_token_configuration() {
-        assert_auth_invalidation_cancels_inflight_configuration(false).await;
-    }
-
-    #[tokio::test]
-    async fn logout_invalidates_and_fails_closed_an_inflight_token_configuration() {
-        assert_auth_invalidation_cancels_inflight_configuration(true).await;
-    }
-
-    async fn spawn_stalled_witness_configuration() -> (
-        tempfile::TempDir,
-        std::sync::Arc<PluginDbRuntime>,
-        MockServer,
-        tokio::task::JoinHandle<Result<crate::CloudsyncTokenConfigurationResult>>,
-    ) {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(4),
-        })
-        .await
-        .unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        sqlx::query(
-            "UPDATE storage_migration_state
-             SET importer_version = ?, parity_verified = 1
-             WHERE id = 'legacy_v1'",
-        )
-        .bind(anlg_db_app::LEGACY_IMPORTER_VERSION)
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(std::sync::Arc::new(db)));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("user-a", &recovery_key)
-            .unwrap();
-        assert!(
-            runtime
-                .bind_cloudsync_account("user-a".to_string())
-                .await
-                .unwrap()
-        );
-        let witness_server = MockServer::start().await;
-        Mock::given(path("/sync/e2ee/witness/user-a"))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .set_delay(std::time::Duration::from_secs(60))
-                    .set_body_json(serde_json::json!({
-                        "initialized": true,
-                        "initializedAt": "2026-07-23T00:00:00Z",
-                        "headSequence": 0,
-                        "throughSequence": 0,
-                        "nextAfterSequence": 0,
-                        "events": [],
-                    })),
-            )
-            .mount(&witness_server)
-            .await;
-        let witness_endpoint = format!("{}/sync/e2ee/witness/user-a", witness_server.uri());
-
-        let configure_runtime = std::sync::Arc::clone(&runtime);
-        let configure = tokio::spawn(async move {
-            configure_runtime
-                .configure_cloudsync_token(
-                    "managed-database-id".to_string(),
-                    "stale-token".to_string(),
-                    "user-a".to_string(),
-                    crate::CloudsyncE2eeWitness {
-                        endpoint: witness_endpoint,
-                        access_token: "access-token".to_string(),
-                    },
-                )
-                .await
-        });
-        tokio::time::timeout(std::time::Duration::from_secs(10), async {
-            loop {
-                if !witness_server
-                    .received_requests()
-                    .await
-                    .expect("failed to inspect witness requests")
-                    .is_empty()
-                {
-                    return;
-                }
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("token configuration did not reach the witness");
-        (dir, runtime, witness_server, configure)
-    }
-
-    #[tokio::test]
-    async fn suspend_cancels_a_stalled_witness_configuration() {
-        let (_dir, runtime, _witness_server, configure) =
-            spawn_stalled_witness_configuration().await;
-        tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            runtime.suspend_cloudsync(),
-        )
-        .await
-        .expect("suspension waited for the stalled witness request")
-        .unwrap();
-        let error = configure.await.unwrap().unwrap_err();
-        assert!(matches!(
-            error,
-            crate::Error::CloudsyncConfigurationCancelled
-        ));
-    }
-
-    #[cfg(target_os = "macos")]
-    #[tokio::test]
-    async fn suspend_drains_token_configuration_during_large_replica_cleanup() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = std::sync::Arc::new(
-            Db::open(DbOpenOptions {
-                storage: DbStorage::Local(&db_path),
-                cloudsync_enabled: true,
-                journal_mode_wal: true,
-                foreign_keys: true,
-                max_connections: Some(4),
-            })
-            .await
-            .unwrap(),
-        );
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        sqlx::query(
-            "UPDATE storage_migration_state
-             SET importer_version = ?, parity_verified = 1
-             WHERE id = 'legacy_v1'",
-        )
-        .bind(anlg_db_app::LEGACY_IMPORTER_VERSION)
-        .execute(db.pool())
-        .await
-        .unwrap();
-        anlg_db_app::claim_cloudsync_workspace(db.pool(), "user-a")
-            .await
-            .unwrap();
-        anlg_db_app::set_cloudsync_personal_write_scope(db.pool(), "user-a")
-            .await
-            .unwrap();
-        sqlx::query(
-            "WITH RECURSIVE rows(id) AS (
-               SELECT 1
-               UNION ALL
-               SELECT id + 1 FROM rows WHERE id < 100000
-             )
-             INSERT INTO e2ee_records (id, workspace_id, payload)
-             SELECT
-               printf('record-%d', id),
-               'user-a',
-               printf('ciphertext-%d', id)
-             FROM rows",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records")
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            100_000
-        );
-        db.cloudsync_init(CLOUDSYNC_REPLICA_TABLE, None, None)
-            .await
-            .unwrap();
-        db.cloudsync_set_filter(CLOUDSYNC_REPLICA_TABLE, CLOUDSYNC_WRITE_FILTER)
-            .await
-            .unwrap();
-
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(std::sync::Arc::clone(&db)));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("user-a", &recovery_key)
-            .unwrap();
-        let key = runtime.workspace_key("user-a").unwrap();
-        let generation =
-            anlg_db_app::stage_cloudsync_poison_recovery(db.pool(), "user-a", "user-a")
-                .await
-                .unwrap();
-        let recovery = anlg_db_app::ensure_cloudsync_recovery_state(
-            db.pool(),
-            &generation,
-            "user-a",
-            "user-a",
-            &key,
-        )
-        .await
-        .unwrap();
-        assert_eq!(
-            recovery.phase,
-            anlg_db_app::CloudsyncRecoveryPhase::NeedFirstLogout
-        );
-
-        let witness_state = InitiallyUninitializedWitness::default();
-        witness_state.initialized.store(true, Ordering::SeqCst);
-        let witness_server = MockServer::start().await;
-        Mock::given(path("/sync/e2ee/witness/user-a"))
-            .respond_with(witness_state)
-            .mount(&witness_server)
-            .await;
-        let witness_endpoint = format!("{}/sync/e2ee/witness/user-a", witness_server.uri());
-
-        let configure_runtime = std::sync::Arc::clone(&runtime);
-        let configure = tokio::spawn(async move {
-            configure_runtime
-                .configure_cloudsync_token(
-                    "managed-database-id".to_string(),
-                    "stale-token".to_string(),
-                    "user-a".to_string(),
-                    crate::CloudsyncE2eeWitness {
-                        endpoint: witness_endpoint,
-                        access_token: "access-token".to_string(),
-                    },
-                )
-                .await
-        });
-        tokio::time::timeout(std::time::Duration::from_secs(10), async {
-            while !db.cloudsync_interrupt_registered() {
-                assert!(
-                    !configure.is_finished(),
-                    "token configuration finished before native replica cleanup"
-                );
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("token configuration did not register native replica cleanup");
-
-        let suspended_at = std::time::Instant::now();
-        tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            runtime.suspend_cloudsync(),
-        )
-        .await
-        .expect("suspension did not drain native replica cleanup")
-        .unwrap();
-        assert!(
-            suspended_at.elapsed() < std::time::Duration::from_secs(1),
-            "suspension exceeded the cancellation deadline"
-        );
-
-        let error = configure.await.unwrap().unwrap_err();
-        assert!(matches!(
-            error,
-            crate::Error::CloudsyncConfigurationCancelled
-        ));
-        assert!(!db.cloudsync_interrupt_registered());
-
-        let idle_at = std::time::Instant::now();
-        db.cloudsync_wait_for_sync_idle().await;
-        assert!(
-            idle_at.elapsed() < std::time::Duration::from_millis(250),
-            "CloudSync worker was not idle after suspension"
-        );
-        db.cloudsync_version()
-            .await
-            .expect("CloudSync SQLite worker crashed during cancellation");
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["configured"], false);
-        assert_eq!(status["running"], false);
-        assert_eq!(status["network_initialized"], false);
-
-        let inserted_at = std::time::Instant::now();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, title)
-             VALUES ('local-after-cleanup-cancellation', 'user-a', 'Local')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            inserted_at.elapsed() < std::time::Duration::from_millis(250),
-            "local write remained blocked after configuration cancellation"
-        );
-    }
-
-    #[tokio::test]
-    async fn activity_begin_cancels_and_drains_a_stalled_witness_configuration() {
-        let (_dir, runtime, _witness_server, configure) =
-            spawn_stalled_witness_configuration().await;
-
-        tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            runtime.begin_cloudsync_activity_with_timeout(
-                "capture".to_string(),
-                "session-1".to_string(),
-                std::time::Duration::from_millis(500),
-            ),
-        )
-        .await
-        .expect("activity begin waited for the stalled witness request")
-        .unwrap();
-        let error = configure.await.unwrap().unwrap_err();
-        assert!(matches!(error, crate::Error::CloudsyncActivityDeferred));
-        assert!(runtime.e2ee_sync_hook.activity_paused());
-        tokio::time::timeout(
-            std::time::Duration::from_millis(250),
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('activity-write', 'user-a', 'user-a', 'Local write')",
-            )
-            .execute(runtime.pool()),
-        )
-        .await
-        .expect("local write remained blocked after configuration drained")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn stale_configuration_waiting_after_a_newer_attempt_does_not_clear_its_key() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(db);
-        let stale_generation = runtime.begin_cloudsync_auth_configuration();
-        let _newer_generation = runtime.begin_cloudsync_auth_configuration();
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("new-user", &recovery_key)
-            .unwrap();
-
-        let error = runtime
-            .configure_cloudsync_token_with_projection_at_generation(
-                CloudsyncTokenConfiguration::new(
-                    "managed-database-id".to_string(),
-                    "stale-token".to_string(),
-                    "old-user".to_string(),
-                    None,
-                    crate::CloudsyncE2eeWitness {
-                        endpoint: "https://example.com/witness".to_string(),
-                        access_token: "access-token".to_string(),
-                    },
-                ),
-                Some(("old-user".to_string(), recovery_key)),
-                stale_generation,
-            )
-            .await
-            .unwrap_err();
-
-        assert!(matches!(
-            error,
-            crate::Error::CloudsyncConfigurationCancelled
-        ));
-        assert!(runtime.workspace_key("new-user").is_some());
-        assert!(runtime.workspace_key("old-user").is_none());
-    }
-
-    #[tokio::test]
-    async fn suspend_joins_full_resync_before_returning() {
-        let runtime = PluginDbRuntime::new(std::sync::Arc::new(
-            Db::connect_memory_plain().await.unwrap(),
-        ));
-        let finished = install_waiting_full_resync_task(&runtime).await;
-
-        runtime.suspend_cloudsync().await.unwrap();
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), finished)
-            .await
-            .unwrap()
-            .unwrap();
-        assert!(runtime.cloudsync_full_resync_task.lock().await.is_none());
-        assert!(
-            !runtime
-                .scheduled_cloudsync_full_resync
-                .lock()
-                .unwrap()
-                .is_active("test-generation")
-        );
-    }
-
-    #[tokio::test]
-    async fn suspend_cancels_recovery_before_waiting_for_its_control_guard() {
-        let runtime = PluginDbRuntime::new(std::sync::Arc::new(
-            Db::connect_memory_plain().await.unwrap(),
-        ));
-        let control_operation = std::sync::Arc::clone(&runtime.cloudsync_control_operation);
-        let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
-        let (control_acquired_tx, control_acquired_rx) = tokio::sync::oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            let _control_operation = control_operation.lock().await;
-            let _ = control_acquired_tx.send(());
-            let _ = shutdown_rx.await;
-        });
-        runtime
-            .scheduled_cloudsync_full_resync
-            .lock()
-            .unwrap()
-            .claim("test-generation");
-        *runtime.cloudsync_full_resync_task.lock().await = Some(CloudsyncFullResyncTask {
-            config: test_full_resync_config("test-token"),
-            generation: "test-generation".to_string(),
-            shutdown_tx: Some(shutdown_tx),
-            join_handle,
-        });
-        tokio::time::timeout(std::time::Duration::from_secs(1), control_acquired_rx)
-            .await
-            .expect("recovery task did not acquire CloudSync control")
-            .unwrap();
-
-        tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            runtime.suspend_cloudsync(),
-        )
-        .await
-        .expect("suspension waited for recovery instead of cancelling it")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn full_resync_cancellation_drains_while_control_is_held() {
-        let runtime = PluginDbRuntime::new(std::sync::Arc::new(
-            Db::connect_memory_plain().await.unwrap(),
-        ));
-        let control = runtime.cloudsync_control_operation.lock().await;
-        runtime
-            .schedule_cloudsync_full_resync(
-                "test-generation".to_string(),
-                test_full_resync_config("test-token"),
-                runtime.cloudsync_auth_generation(),
-            )
-            .await;
-        tokio::time::timeout(std::time::Duration::from_secs(1), async {
-            while runtime
-                .e2ee_sync_hook
-                .full_resync_control_waits
-                .load(Ordering::Acquire)
-                == 0
-            {
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("full resync did not wait for CloudSync control");
-
-        tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            runtime.cancel_cloudsync_full_resync(),
-        )
-        .await
-        .expect("full resync cancellation deadlocked behind CloudSync control");
-        assert!(runtime.cloudsync_full_resync_task.lock().await.is_none());
-        drop(control);
-    }
-
-    #[tokio::test]
-    async fn logout_joins_full_resync_before_returning() {
-        let runtime = PluginDbRuntime::new(std::sync::Arc::new(
-            Db::connect_memory_plain().await.unwrap(),
-        ));
-        let finished = install_waiting_full_resync_task(&runtime).await;
-
-        runtime.logout_cloudsync(false).await.unwrap();
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), finished)
-            .await
-            .unwrap()
-            .unwrap();
-        assert!(runtime.cloudsync_full_resync_task.lock().await.is_none());
-    }
-
-    #[tokio::test]
-    async fn dropping_runtime_signals_full_resync_task_shutdown() {
-        let runtime = PluginDbRuntime::new(std::sync::Arc::new(
-            Db::connect_memory_plain().await.unwrap(),
-        ));
-        let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
-        let (started_tx, started_rx) = tokio::sync::oneshot::channel();
-        let (shutdown_observed_tx, shutdown_observed_rx) = tokio::sync::oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            let _ = started_tx.send(());
-            shutdown_rx.await.unwrap();
-            let _ = shutdown_observed_tx.send(());
-        });
-        *runtime.cloudsync_full_resync_task.lock().await = Some(CloudsyncFullResyncTask {
-            config: test_full_resync_config("test-token"),
-            generation: "test-generation".to_string(),
-            shutdown_tx: Some(shutdown_tx),
-            join_handle,
-        });
-        started_rx.await.unwrap();
-
-        drop(runtime);
-
-        tokio::time::timeout(std::time::Duration::from_secs(1), shutdown_observed_rx)
-            .await
-            .unwrap()
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn poisoned_replica_recovery_requires_the_disposable_e2ee_table_only() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        anlg_db_app::claim_cloudsync_workspace(db.pool(), "workspace-1")
-            .await
-            .unwrap();
-        sqlx::query(
-            "CREATE TABLE cloudsync_table_settings (
-               tbl_name TEXT NOT NULL,
-               col_name TEXT NOT NULL,
-               key TEXT NOT NULL,
-               value TEXT,
-               PRIMARY KEY (tbl_name, col_name, key)
-             )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO cloudsync_table_settings (tbl_name, col_name, key, value)
-             VALUES ('e2ee_records', '*', 'filter', 'workspace_id')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-
-        require_disposable_cloudsync_replica(
-            runtime.db.as_ref(),
-            CloudsyncOperationCancellation::None,
-        )
-        .await
-        .unwrap();
-
-        sqlx::query(
-            "INSERT INTO cloudsync_table_settings (tbl_name, col_name, key, value)
-             VALUES ('sessions', '*', 'filter', 'workspace_id')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert!(
-            require_disposable_cloudsync_replica(
-                runtime.db.as_ref(),
-                CloudsyncOperationCancellation::None,
-            )
-            .await
-            .unwrap_err()
-            .to_string()
-            .contains("sessions")
-        );
-        let key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap()
-        .workspace_key("workspace-1")
-        .unwrap();
-        assert!(
-            prepare_cloudsync_poison_recovery(
-                runtime.db.as_ref(),
-                "workspace-1",
-                "workspace-1",
-                &key,
-                CloudsyncOperationCancellation::None,
-            )
-            .await
-            .unwrap_err()
-            .to_string()
-            .contains("sessions")
-        );
-        assert_eq!(
-            anlg_db_app::cloudsync_full_resync_generation(runtime.db.pool())
-                .await
-                .unwrap(),
-            None
-        );
-        assert!(
-            anlg_db_app::cloudsync_recovery_state(runtime.db.pool())
-                .await
-                .unwrap()
-                .is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn native_replica_logout_preserves_domain_rows_and_reinstalls_an_empty_filter() {
-        let db = Db::connect_memory().await.unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        anlg_db_app::set_cloudsync_personal_write_scope(db.pool(), "workspace-1")
-            .await
-            .unwrap();
-        db.cloudsync_init(CLOUDSYNC_REPLICA_TABLE, None, None)
-            .await
-            .unwrap();
-        db.cloudsync_set_filter(CLOUDSYNC_REPLICA_TABLE, CLOUDSYNC_WRITE_FILTER)
-            .await
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, title)
-             VALUES ('session-1', 'workspace-1', 'Preserved')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_records (id, workspace_id, payload)
-             VALUES ('record-1', 'workspace-1', 'ciphertext')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_local_state (
-               record_id, workspace_id, table_name, row_id, field_name,
-               revision, value_tag, payload_hash, writer_id, payload
-             ) VALUES (
-               'record-1', 'workspace-1', 'sessions', 'session-1', 'title',
-               1, 'value-tag', 'payload-hash', 'writer-1', 'ciphertext'
-             )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let local_changes_before: i64 =
-            sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_records_cloudsync")
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert!(local_changes_before > 0);
-        let dirty_rows_before: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert!(dirty_rows_before > 0);
-        let local_state_before: (String, String) = sqlx::query_as(
-            "SELECT payload_hash, payload
-             FROM e2ee_local_state
-             WHERE record_id = 'record-1'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-
-        db.cloudsync_network_logout().await.unwrap();
-
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM sessions")
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            1
-        );
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records")
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            0
-        );
-        db.cloudsync_init(CLOUDSYNC_REPLICA_TABLE, None, None)
-            .await
-            .unwrap();
-        db.cloudsync_set_filter(CLOUDSYNC_REPLICA_TABLE, CLOUDSYNC_WRITE_FILTER)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records_cloudsync")
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            0
-        );
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>(
-                "SELECT COUNT(*)
-                 FROM cloudsync_table_settings
-                 WHERE tbl_name = 'e2ee_records'
-                   AND col_name = '*'
-                   AND key = 'filter'
-                   AND value = ?",
-            )
-            .bind(CLOUDSYNC_WRITE_FILTER)
-            .fetch_one(db.pool())
-            .await
-            .unwrap(),
-            1
-        );
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_dirty_rows")
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            dirty_rows_before
-        );
-        assert_eq!(
-            sqlx::query_as::<_, (String, String)>(
-                "SELECT payload_hash, payload
-                 FROM e2ee_local_state
-                 WHERE record_id = 'record-1'",
-            )
-            .fetch_one(db.pool())
-            .await
-            .unwrap(),
-            local_state_before
-        );
-    }
-
-    #[tokio::test]
-    async fn legacy_cutover_snapshots_local_state_before_initializing_the_witness() {
-        let db = std::sync::Arc::new(Db::connect_memory().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        db.cloudsync_init("sessions", None, None).await.unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-1', 'workspace-1', 'Legacy session')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_local_state")
-                .fetch_one(db.pool())
-                .await
-                .unwrap(),
-            0
-        );
-
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        let workspace_key = runtime.workspace_key("workspace-1").unwrap();
-        let witness_state = InitiallyUninitializedWitness::default();
-        let witness_server = MockServer::start().await;
-        Mock::given(path("/sync/e2ee/witness/workspace-1"))
-            .respond_with(witness_state.clone())
-            .mount(&witness_server)
-            .await;
-        let witness = crate::e2ee_witness::E2eeWitnessClient::new(
-            crate::CloudsyncE2eeWitness {
-                endpoint: format!("{}/sync/e2ee/witness/workspace-1", witness_server.uri()),
-                access_token: "access-token".to_string(),
-            },
-            "workspace-1",
-        )
-        .unwrap();
-
-        let cancellation = crate::e2ee_witness::E2eeWitnessCancellation::default();
-        runtime
-            .prepare_e2ee_cutover_and_initialize_witness(&witness, &workspace_key, &cancellation)
-            .await
-            .unwrap();
-
-        assert!(witness_state.initialized.load(Ordering::SeqCst));
-        assert!(
-            anlg_db_app::has_e2ee_local_state(db.pool(), "workspace-1")
-                .await
-                .unwrap()
-        );
-        assert!(
-            !anlg_db_core::cloudsync_is_enabled_on(db.pool(), "sessions")
-                .await
-                .unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn reconciliation_barrier_blocks_renderer_writes() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(db));
-        let guard = runtime.synced_write_barrier.write().await;
-
-        let execute_runtime = std::sync::Arc::clone(&runtime);
-        let mut execute = tokio::spawn(async move {
-            execute_runtime
-                .execute(
-                    "INSERT INTO sessions (id, title) VALUES ('session-1', 'Session 1')"
-                        .to_string(),
-                    vec![],
-                )
-                .await
-        });
-        let transaction_runtime = std::sync::Arc::clone(&runtime);
-        let mut transaction = tokio::spawn(async move {
-            transaction_runtime
-                .execute_transaction(vec![TransactionStatement {
-                    sql: "INSERT INTO sessions (id, title) VALUES ('session-2', 'Session 2')"
-                        .to_string(),
-                    params: vec![],
-                    expected_rows_affected: Some(1),
-                }])
-                .await
-        });
-        let proxy_runtime = std::sync::Arc::clone(&runtime);
-        let mut proxy = tokio::spawn(async move {
-            proxy_runtime
-                .execute_proxy(
-                    "INSERT INTO sessions (id, title) VALUES ('session-3', 'Session 3')"
-                        .to_string(),
-                    vec![],
-                    ProxyQueryMethod::Run,
-                )
-                .await
-        });
-
-        let timeout = std::time::Duration::from_millis(25);
-        assert!(tokio::time::timeout(timeout, &mut execute).await.is_err());
-        assert!(
-            tokio::time::timeout(timeout, &mut transaction)
-                .await
-                .is_err()
-        );
-        assert!(tokio::time::timeout(timeout, &mut proxy).await.is_err());
-        drop(guard);
-
-        execute.await.unwrap().unwrap();
-        transaction.await.unwrap().unwrap();
-        proxy.await.unwrap().unwrap();
-
-        let session_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM sessions")
-            .fetch_one(runtime.pool())
-            .await
-            .unwrap();
-        assert_eq!(session_count, 3);
-    }
-
-    #[tokio::test]
-    async fn reconciliation_barrier_blocks_native_synced_writes() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        let runtime = std::sync::Arc::new(PluginDbRuntime::new(db));
-        let guard = runtime.synced_write_barrier.write().await;
-        let write_runtime = std::sync::Arc::clone(&runtime);
-        let mut write = tokio::spawn(async move {
-            let _guard = write_runtime.synced_write_guard().await;
-        });
-
-        assert!(
-            tokio::time::timeout(std::time::Duration::from_millis(25), &mut write)
-                .await
-                .is_err()
-        );
-        drop(guard);
-        write.await.unwrap();
-    }
-
-    fn unavailable_extension_error() -> DbOpenError {
-        DbOpenError::Io(std::io::Error::other("cloudsync extension unavailable"))
-    }
-
-    fn failed_extension_probe_error() -> DbOpenError {
-        DbOpenError::Io(std::io::Error::other("cloudsync extension probe failed"))
-    }
-
-    #[tokio::test]
-    async fn cloudsync_open_failure_falls_back_for_uninitialized_database() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-
-        let db = open_app_db_without_cloudsync(
-            DbStorage::Local(&db_path),
-            unavailable_extension_error(),
-            failed_extension_probe_error(),
-        )
-        .await
-        .unwrap();
-
-        assert!(!db.cloudsync_enabled());
-        let sessions_exists: bool = sqlx::query_scalar(
-            "SELECT EXISTS(
-                SELECT 1 FROM sqlite_master WHERE type = 'table' AND name = 'sessions'
-            )",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert!(sessions_exists);
-    }
-
-    #[cfg(any(
-        all(target_os = "macos", target_arch = "aarch64"),
-        all(target_os = "macos", target_arch = "x86_64"),
-        all(target_os = "linux", target_env = "gnu", target_arch = "aarch64"),
-        all(target_os = "linux", target_env = "gnu", target_arch = "x86_64"),
-        all(target_os = "linux", target_env = "musl", target_arch = "aarch64"),
-        all(target_os = "linux", target_env = "musl", target_arch = "x86_64"),
-        all(target_os = "windows", target_arch = "x86_64"),
-    ))]
-    #[tokio::test]
-    async fn extension_open_without_initialized_tables_allows_plain_fallback() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(1),
-        })
-        .await
-        .unwrap();
-        db.pool().close().await;
-        drop(db);
-
-        let db = open_app_db_without_cloudsync(
-            DbStorage::Local(&db_path),
-            unavailable_extension_error(),
-            failed_extension_probe_error(),
-        )
-        .await
-        .unwrap();
-
-        assert!(!db.cloudsync_enabled());
-        assert!(!database_uses_cloudsync_schema(&db).await.unwrap());
-    }
-
-    #[cfg(any(
-        all(target_os = "macos", target_arch = "aarch64"),
-        all(target_os = "macos", target_arch = "x86_64"),
-        all(target_os = "linux", target_env = "gnu", target_arch = "aarch64"),
-        all(target_os = "linux", target_env = "gnu", target_arch = "x86_64"),
-        all(target_os = "linux", target_env = "musl", target_arch = "aarch64"),
-        all(target_os = "linux", target_env = "musl", target_arch = "x86_64"),
-        all(target_os = "windows", target_arch = "x86_64"),
-    ))]
-    #[tokio::test]
-    async fn cloudsync_open_failure_does_not_migrate_initialized_replica_plainly() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(1),
-        })
-        .await
-        .unwrap();
-        sqlx::query(
-            "CREATE TABLE items (
-                id TEXT PRIMARY KEY NOT NULL,
-                value TEXT NOT NULL DEFAULT ''
-            )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.cloudsync_init("items", None, None).await.unwrap();
-        db.pool().close().await;
-        drop(db);
-
-        let error = open_app_db_without_cloudsync(
-            DbStorage::Local(&db_path),
-            unavailable_extension_error(),
-            failed_extension_probe_error(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(matches!(error, crate::Error::Db(DbOpenError::Io(_))));
-        let plain = Db::connect_local_plain(&db_path).await.unwrap();
-        let sessions_exists: bool = sqlx::query_scalar(
-            "SELECT EXISTS(
-                SELECT 1 FROM sqlite_master WHERE type = 'table' AND name = 'sessions'
-            )",
-        )
-        .fetch_one(plain.pool())
-        .await
-        .unwrap();
-        assert!(!sessions_exists);
-    }
-
-    #[tokio::test]
-    async fn cloudsync_open_fallback_propagates_schema_errors() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = Db::open(app_db_open_options(DbStorage::Local(&db_path), false))
-            .await
-            .unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = 'not-json'
-             WHERE id = 'cloudsync_workspace_binding'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.pool().close().await;
-        drop(db);
-
-        let error = open_app_db_without_cloudsync(
-            DbStorage::Local(&db_path),
-            unavailable_extension_error(),
-            failed_extension_probe_error(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(matches!(
-            error,
-            crate::Error::AppSchema(anlg_db_app::AppSchemaError::CloudsyncWorkspace(
-                anlg_db_app::CloudsyncWorkspaceError::InvalidBinding
-            ))
-        ));
-    }
-
-    #[tokio::test]
-    async fn failed_cloudsync_preflight_clears_new_credentials() {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(4),
-        })
-        .await
-        .unwrap();
-        anlg_db_app::prepare_schema(&db).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::new(db));
-        let cancellation = crate::e2ee_witness::E2eeWitnessCancellation::default();
-
-        runtime
-            .prepare_cloudsync_config_fail_closed(
-                anlg_db_core::CloudsyncRuntimeConfig {
-                    connection_string: "managed-database-id".to_string(),
-                    auth: anlg_db_core::CloudsyncAuth::Token {
-                        token: "secret-token".to_string(),
-                    },
-                    tables: vec![anlg_db_core::CloudsyncTableSpec {
-                        table_name: "missing_table".to_string(),
-                        crdt_algo: None,
-                        init_flags: None,
-                        enabled: true,
-                    }],
-                    sync_interval_ms: DEFAULT_CLOUDSYNC_INTERVAL_MS,
-                    wait_ms: Some(5_000),
-                    max_retries: Some(3),
-                },
-                &cancellation,
-            )
-            .await
-            .unwrap_err();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["configured"], false);
-        assert_eq!(status["running"], false);
-        assert_eq!(status["network_initialized"], false);
-    }
-
-    #[tokio::test]
-    async fn cloudsync_status_reports_pending_e2ee_dirty_rows() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, title)
-             VALUES ('session-1', 'workspace-1', 'Local edit')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-
-        assert_eq!(status["has_unsent_changes"], true);
-    }
-
-    #[tokio::test]
-    async fn cloudsync_status_does_not_treat_inbound_reconciliation_as_unsent() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_witness_records (
-               workspace_id,
-               record_id,
-               revision,
-               writer_id,
-               payload_hash,
-               payload,
-               sequence
-             )
-             VALUES ('workspace-1', 'record-1', 1, 'writer-1', 'hash', 'payload', 1)",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-
-        assert!(runtime.e2ee_sync_hook.reconciliation_requested());
-        assert_ne!(status["has_unsent_changes"], true);
-    }
-
-    #[tokio::test]
-    async fn cloudsync_status_ignores_only_the_active_transcript() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        let recovery_key = anlg_e2ee::RecoveryKey::parse(
-            "anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc",
-        )
-        .unwrap();
-        runtime
-            .set_e2ee_recovery_key("workspace-1", &recovery_key)
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO transcripts (id, workspace_id, session_id, words_json)
-             VALUES ('transcript-1', 'workspace-1', 'session-1', '[{\"text\":\"partial\"}]')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO app_settings (id, value_json)
-             VALUES ('capture_lifecycle_pending:session-1', ?)",
-        )
-        .bind(
-            serde_json::json!({
-                "version": 1,
-                "phase": "capturing",
-                "sessionId": "session-1",
-                "transcriptId": "transcript-1",
-                "startedAt": 1_000,
-                "createdAt": "2026-07-24T00:00:00.000Z",
-                "audioOffsetMs": 0,
-                "preserveExistingTranscript": false,
-                "ownerUserId": "workspace-1",
-                "memo": ""
-            })
-            .to_string(),
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_ne!(status["has_unsent_changes"], true);
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(value_json, '$.phase', 'finalizing')
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let status = runtime.cloudsync_status().await.unwrap();
-        assert_eq!(status["has_unsent_changes"], true);
-    }
-
-    #[tokio::test]
-    async fn repeated_cloudsync_status_polling_returns_base_status_on_a_busy_pool() {
-        let db = std::sync::Arc::new(Db::connect_memory_plain().await.unwrap());
-        anlg_db_app::prepare_schema(db.as_ref()).await.unwrap();
-        let runtime = PluginDbRuntime::new(std::sync::Arc::clone(&db));
-        let mut held_connection = db.pool().acquire().await.unwrap();
-        let started = std::time::Instant::now();
-
-        for _ in 0..32 {
-            let status = tokio::time::timeout(
-                std::time::Duration::from_millis(50),
-                runtime.cloudsync_status(),
-            )
-            .await
-            .expect("CloudSync status waited for a busy pool")
-            .unwrap();
-            assert_eq!(status["activity_paused"], false);
-            assert_eq!(status["has_unsent_changes"], serde_json::Value::Null);
-            assert!(status.get("recovery_pending").is_none());
-        }
-        assert!(started.elapsed() < std::time::Duration::from_secs(1));
-
-        held_connection.return_to_pool().await;
-        for _ in 0..32 {
-            tokio::time::timeout(
-                std::time::Duration::from_millis(100),
-                runtime.cloudsync_status(),
-            )
-            .await
-            .expect("CloudSync status left SQLite work in flight")
-            .unwrap();
-        }
-        assert_eq!(db.pool().num_idle(), 1);
-        tokio::time::timeout(std::time::Duration::from_millis(100), db.pool().acquire())
-            .await
-            .expect("repeated CloudSync status polling exhausted the pool")
-            .unwrap();
-    }
-}
+mod tests;
